fix: remove redundant cast

+++ src/mlcg/neighbor_list/nvalchemi_impl.py
@@ -100,7 +100,7 @@ def nvalchemi_naive_neighbor_list(
         (idx_i, idx_j), _, idx_S = result
         # cell_shifts = torch.einsum("ni,nij->nj", idx_S.to(cell.dtype), cell[data.batch[idx_i]])
         cell_shifts = (
-            idx_S.to(cell.dtype).to(cell.dtype).unsqueeze(-1)
+            idx_S.to(cell.dtype).unsqueeze(-1)
             * cell[data.batch[idx_i]]
         ).sum(dim=1)
         return idx_i, idx_j, cell_shifts, None
@@ -197,7 +197,7 @@ def nvalchemi_cell_neighbor_list(
     if with_pbc:
         # cell_shifts = torch.einsum("ni,nij->nj", idx_S.to(cell.dtype), cell[data.batch[idx_i]])
         cell_shifts = (
-            idx_S.to(cell.dtype).to(cell.dtype).unsqueeze(-1)
+            idx_S.to(cell.dtype).unsqueeze(-1)
             * cell[data.batch[idx_i]]
         ).sum(dim=1)
     else:
